backend/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12,<4.0
psycopg2>2.8
celery>=5.2
requests>=2.26
wikibaseintegrator~=0.11.0
django-cors-headers==3.10.1
